+++ attendance-manager/src/attendanceManager.ts
