+++ attendance-manager/src/auth.ts
