+++ attendance-manager/src/command.ts
