+++ attendance-manager/src/config.ts
