+++ attendance-manager/src/index.ts
@@ -2,7 +2,7 @@
  * @file GASエディタから実行できる関数を定義する
  */
 
-import { periodicallyCheckForAttendanceManager, initAttendanceManager } from "./attendance-manager/attendanceManager";
+import { periodicallyCheckForAttendanceManager, initAttendanceManager } from "./attendanceManager";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 declare const global: any;
