+++ attendance-manager/src/message.test.ts
