+++ attendance-manager/src/message.ts
