+++ attendance-manager/src/reaction.ts
