+++ attendance-manager/src/utilities.ts
